+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 ### FEATURES:
 * builder/docker: add packer docker builder support
 * builder/file: add packer file builder support
+* provisioner/DockerImport: add packer docker import provisioner support
 * provisioner/file: add packer file provisioner support